+++ Sources/KeyboardKit/Styles/SecondaryInputCalloutStyle.swift
@@ -17,31 +17,60 @@ import SwiftUI
  */
 public struct SecondaryInputCalloutStyle {
     
+    /**
+     Create a secondary input callout style.
+     
+     - Parameters:
+       - callout: The callout style to use.
+       - font: The font to use in the callout.
+       - selectedBackgroundColor: The background color of the selected item.
+       - selectedForegroundColor: The foreground color of the selected item.
+       - verticalTextPadding: The vertical padding to apply to text in the callout.
+     */
     public init(
         callout: CalloutStyle = .standard,
         font: Font = Self.standardFont,
         selectedBackgroundColor: Color = Color.blue,
-        selectedTextColor: Color = Color.white,
-        verticalPadding: CGFloat = 5) {
+        selectedForegroundColor: Color = Color.white,
+        verticalTextPadding: CGFloat = 5) {
         self.callout = callout
         self.font = font
         self.selectedBackgroundColor = selectedBackgroundColor
-        self.selectedTextColor = selectedTextColor
-        self.verticalPadding = verticalPadding
+        self.selectedForegroundColor = selectedForegroundColor
+        self.verticalTextPadding = verticalTextPadding
     }
     
+    /**
+     The callout style to use.
+     */
     public var callout: CalloutStyle
+    
+    /**
+     The font to use in the callout.
+     */
     public var font: Font
+    
+    /**
+     The background color of the selected item.
+     */
     public var selectedBackgroundColor: Color
-    public var selectedTextColor: Color
-    public var verticalPadding: CGFloat
+    
+    /**
+     The foreground color of the selected item.
+     */
+    public var selectedForegroundColor: Color
+    
+    /**
+     The vertical padding to apply to text in the callout.
+     */
+    public var verticalTextPadding: CGFloat
 }
 
 public extension SecondaryInputCalloutStyle {
     
     /**
-     This is the standard callout style that will be used by
-     default. It looks like a system keyboard callout.
+     This standard style will be used by default. It aims to
+     look like a native system keyboard's secondary callout.
      */
     static var standard = SecondaryInputCalloutStyle()