+++ Sources/KeyboardKit/Styles/SystemKeyboardButtonBorderStyle.swift
@@ -14,6 +14,13 @@ import SwiftUI
  */
 public struct SystemKeyboardButtonBorderStyle {
     
+    /**
+     Create a system keyboard button border style.
+     
+     - Parameters:
+       - color: The color of the border.
+       - size: The size of the border.
+     */
     public init(
         color: Color,
         size: CGFloat) {
@@ -21,7 +28,14 @@ public struct SystemKeyboardButtonBorderStyle {
         self.size = size
     }
     
+    /**
+     The color of the border.
+     */
     public var color: Color
+    
+    /**
+     The size of the border.
+     */
     public var size: CGFloat
 }
 
@@ -29,9 +43,15 @@ public extension SystemKeyboardButtonBorderStyle {
     
     /**
      This style means that no border should be applied. This
-     is stye default style for system buttons.
+     is the default style for system buttons.
      */
     static var noBorder: SystemKeyboardButtonBorderStyle {
         SystemKeyboardButtonBorderStyle(color: .clear, size: 0)
     }
+    
+    /**
+     This standard style will be used by default. It aims to
+     look like the border of a native system keyboard button.
+     */
+    static var standard = noBorder
 }